Split up the files properly.
