+++ Lib/test/test_tools/test_c_analyzer/test_c_statics/test_scan.py
@@ -2,7 +2,7 @@ import unittest
 
 from .. import tool_imports_for_tests
 with tool_imports_for_tests():
-    from c_statics import info
+    from c_parser import info
     from c_statics.scan import iter_statics
